Pickeat 58 connect restaurant

See merge request cpelyon/5eti-projet-2024-2025/groupe-7/pick-eat!21

+++ restaurant/src/main/java/fr/pick_eat/restaurant/entity/RestaurantModel.java
@@ -92,7 +92,7 @@ public class RestaurantModel {
         this.GeneratedPicture = GeneratedPicture;
     }
 
-    public double getLat() {
+    public float getLat() {
         return lat;
     }
 
@@ -100,7 +100,7 @@ public class RestaurantModel {
         this.lat = lat;
     }
 
-    public double getLon() {
+    public float getLon() {
         return lon;
     }
